+++ src/Symfony/Bridge/Monolog/Tests/Handler/MailerHandlerTest.php
@@ -14,6 +14,7 @@ namespace Symfony\Bridge\Monolog\Tests\Handler;
 use Monolog\Formatter\HtmlFormatter;
 use Monolog\Formatter\LineFormatter;
 use Monolog\Level;
+use Monolog\Logger;
 use Monolog\LogRecord;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;