+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Compiler/EnableLoggerDebugModePass.php
@@ -11,10 +11,15 @@
 
 namespace Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler;
 
+trigger_deprecation('symfony/framework-bundle', '6.4', 'The "%s" class is deprecated, use argument $debug of HttpKernel\'s Logger instead.', EnableLoggerDebugModePass::class);
+
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Log\Logger;
 
+/**
+ * @deprecated since Symfony 6.4, use argument $debug of HttpKernel's Logger instead
+ */
 final class EnableLoggerDebugModePass implements CompilerPassInterface
 {
     public function process(ContainerBuilder $container): void
@@ -32,7 +37,7 @@ final class EnableLoggerDebugModePass implements CompilerPassInterface
 
     public static function configureLogger(Logger $logger): void
     {
-        if (!\in_array(\PHP_SAPI, ['cli', 'phpdbg'], true) && method_exists($logger, 'enableDebug')) {
+        if (!\in_array(\PHP_SAPI, ['cli', 'phpdbg'], true)) {
             $logger->enableDebug();
         }
     }